% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{add_pct_of_pool_to_samplesheet}
\alias{add_pct_of_pool_to_samplesheet}
\title{Utility function to add percentage of pool to the sample sheet}
\usage{
add_pct_of_pool_to_samplesheet(sample_sheet, object)
}
\arguments{
\item{sample_sheet}{A data frame containing sample metadata,
including \code{sample_alias} and \code{pool} for a hashed sample.}

\item{object}{A Seurat object containing the sample metadata
in its \code{meta.data} slot.}
}
\value{
A modified sample sheet with an additional column \code{pool_fraction}.
}
\description{
Adds a column to the sample sheet indicating how much of the
pool each sample makes up.
}
