% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{annotate_cells}
\alias{annotate_cells}
\title{Annotate cells in a Seurat object using a reference dataset}
\usage{
annotate_cells(
  object,
  reference,
  summarize_by_column = NULL,
  reference_assay = "PNA",
  query_assay = "PNA",
  normalization_method = "LogNormalize",
  reduction = "cca",
  ...
)
}
\arguments{
\item{object}{A Seurat object containing the query data.}

\item{reference}{A Seurat object containing the reference data.}

\item{summarize_by_column}{Optional. A column name in the object metadata to summarize annotations by.}

\item{reference_assay}{The assay in the reference object to use for annotation. Default is "PNA".}

\item{query_assay}{The assay in the query object to use for annotation. Default is "PNA".}

\item{normalization_method}{The normalization method to use. Default is "CLR".}

\item{reduction}{The reduction method to use for finding transfer anchors. Default is "cca".}

\item{...}{Additional arguments passed to \code{FindTransferAnchors} and \code{TransferData}.}
}
\value{
A Seurat object with cell annotations added to the metadata.
}
\description{
This function annotates cells in a Seurat object using a reference dataset.
}
