% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{beaut_label}
\alias{beaut_label}
\title{Beautify axis labels}
\usage{
beaut_label(labels)
}
\arguments{
\item{labels}{A character vector of axis labels.}
}
\value{
A character vector of beautified axis labels.
}
\description{
This function beautifies axis labels by removing underscores and replacing them with spaces.
}
