% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{cluster_palette}
\alias{cluster_palette}
\title{Cluster palette}
\format{
An object of class \code{character} of length 26.
}
\usage{
cluster_palette
}
\description{
A palette for clusters, combining various hues and levels from the PixelgenAccentColors.
}
\keyword{datasets}
