% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{compact_num}
\alias{compact_num}
\title{Compact number formatting}
\usage{
compact_num(n, accuracy = 0.1)
}
\arguments{
\item{n}{A numeric value to be formatted.}

\item{accuracy}{A numeric value specifying the accuracy of the formatting.}
}
\value{
A character string representing the formatted number with the specified size unit.
}
\description{
This function formats a numeric value into a compact representation, either in thousands (k)
or millions (M), with a specified number of decimal places.
}
