% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_abundance_per_celltype}
\alias{component_abundance_per_celltype}
\title{Create the component for abundance per celltype}
\usage{
component_abundance_per_celltype(
  object,
  params,
  sample_palette,
  test_mode = FALSE
)
}
\arguments{
\item{object}{A processed data object containing marker abundance data.}

\item{params}{A list of parameters containing control markers.}

\item{sample_palette}{A color palette for the samples.}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the abundance of each marker
across different celltypes
}
