% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_abundance_per_sample}
\alias{component_abundance_per_sample}
\title{Create the component for abundance per sample}
\usage{
component_abundance_per_sample(
  object,
  params,
  sample_palette,
  sample_levels = NULL,
  test_mode = FALSE
)
}
\arguments{
\item{object}{A processed data object containing marker abundance data.}

\item{params}{A list of parameters, including control markers.}

\item{sample_palette}{A color palette for the samples.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the abundance of each marker
across different samples.
}
