% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_bleedover_noise}
\alias{component_bleedover_noise}
\title{Create the component for bleedover noise}
\usage{
component_bleedover_noise(qc_metrics_tables)
}
\arguments{
\item{qc_metrics_tables}{A list containing a data frame named \code{denoising}}
}
\value{
A ggplot object visualizing the fraction of UMIs removed by denoising.
}
\description{
This function creates a ggplot object that visualizes the fraction of UMIs
removed by denoising.
}
