% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_cell_recovery}
\alias{component_cell_recovery}
\title{Create the component cell recovery}
\usage{
component_cell_recovery(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing processed PXL data, raw QC
metrics, and sample aliases.}
}
\value{
A list containing a plot and a table summarizing the cell recovery.
}
\description{
This function creates a component that visualizes the cell recovery
after filtering components based on size.
}
