% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_clustering_summary}
\alias{component_clustering_summary}
\title{Create a clustering summary component}
\usage{
component_clustering_summary(es_data, heatmap_gradient, test_mode = FALSE)
}
\arguments{
\item{es_data}{An \code{es_data} object containing proximity scores.}

\item{heatmap_gradient}{A color gradient for the heatmap.}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list of ggplot objects summarizing the clustering scores.
}
\description{
This function generates summary plots of the clustering scores.
}
