% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_crossing_edges}
\alias{component_crossing_edges}
\title{Create the component crossing edges}
\usage{
component_crossing_edges(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing QC metrics and sample aliases.}
}
\value{
A list containing a plot and a table summarizing the crossing edges.
}
\description{
This function creates a component that visualizes the percentage of crossing edges
for each sample, both in the initial and refinement stages.
}
