% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_dimred_plots}
\alias{component_dimred_plots}
\title{Create the component dimensionality reduction plots}
\usage{
component_dimred_plots(
  es_data,
  plot_reductions = NULL,
  sample_palette = NULL,
  cluster_palette = NULL,
  sample_plots = FALSE,
  plot_height = 5
)
}
\arguments{
\item{es_data}{An \code{es_data} object containing processed PXL data.}

\item{plot_reductions}{An optional character vector specifying which
dimensionality reductions to plot. By default, all available reductions
are plotted in the preferred order.}

\item{sample_palette}{Optional color palette for samples.}

\item{cluster_palette}{Optional color palette for clusters.}

\item{sample_plots}{Logical indicating whether to include sample-wise plots.}

\item{plot_height}{Numeric value specifying the height of the full plot. Used to space the legend.}
}
\value{
A list containing three plots for reductions.
}
\description{
This function creates a component that visualizes the dimensionality reduction plots
for PCA, colored by clusters, samples, and conditions.
}
