% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_hashing}
\alias{component_hashing}
\title{Create the component hashing}
\usage{
component_hashing(es_data, colors)
}
\arguments{
\item{es_data}{An \code{es_data} object containing sample hashing QC metrics.}

\item{colors}{A gradient color palette to use.}
}
\value{
List with \code{plot}, \code{table}, \code{heatmap_plots_hash_purity}, and \code{heatmap_plots_hash_fraction}.
}
\description{
Create the component hashing
}
