% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_molecule_rank_plot}
\alias{component_molecule_rank_plot}
\title{Create the component for molecule rank plots}
\usage{
component_molecule_rank_plot(pg_data, params)
}
\arguments{
\item{pg_data}{A Seurat object containing the data to be plotted.}

\item{params}{A list of parameters, including \code{molecule_rank_cutoff}, which
is used to draw a horizontal line in the plots.}
}
\value{
A list of ggplot objects, each corresponding to a sample's molecule
rank plot.
}
\description{
This function generates a list of ggplot objects, each representing the
molecule rank plot for a specific sample in the provided Seurat object.
}
