% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_node_degree}
\alias{component_node_degree}
\title{Create the component node degree}
\usage{
component_node_degree(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing processed PXL data and sample aliases.}
}
\value{
A list containing a plot and a table summarizing the mean node degree.
}
\description{
This function creates a component that visualizes the mean node degree for A and B nodes
for each sample.
}
