% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_node_edge_count}
\alias{component_node_edge_count}
\title{Create the component node and edge count}
\usage{
component_node_edge_count(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing raw QC metrics and sample aliases.}
}
\value{
A list containing two plots and a table summarizing the node and edge counts.
}
\description{
This function creates a component that visualizes the number of nodes and edges
before and after cell recovery.
}
