% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_heatmap_sample}
\alias{component_proximity_heatmap_sample}
\title{Create the component for proximity heatmaps per sample and condition}
\usage{
component_proximity_heatmap_sample(
  proximity_scores,
  heatmap_gradient,
  n_markers = 40,
  min_pct_detected = 0.25,
  plot_markers = NULL,
  test_mode = FALSE
)
}
\arguments{
\item{proximity_scores}{A data frame containing proximity scores for different markers.}

\item{heatmap_gradient}{A color palette for the heatmaps.}

\item{n_markers}{The number of markers to plot (default is 40).}

\item{min_pct_detected}{Minimum percentage of cells in which a marker must be detected to be included
(default is 0.25).}

\item{plot_markers}{A vector of markers to plot (default is NULL).}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the proximity scores as heatmaps per sample and condition.
}
