% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_qc_molecule_rank_plot}
\alias{component_qc_molecule_rank_plot}
\title{Create the component for molecule rank plots from qc metrics list}
\usage{
component_qc_molecule_rank_plot(sample_qc_metrics, sample_levels = NULL)
}
\arguments{
\item{sample_qc_metrics}{A list of sample QC metrics.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list of ggplot objects, each corresponding to a sample's molecule
rank plot.
}
\description{
This function generates a list of ggplot objects, each representing the
molecule rank plot for a specific sample in the provided Seurat object.
}
