% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_reads_and_molecules}
\alias{component_sequencing_reads_and_molecules}
\title{Create the component sequencing reads and molecules component}
\usage{
component_sequencing_reads_and_molecules(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing raw QC metrics and sample aliases.}
}
\value{
A list containing plots and a table.
}
\description{
This function creates a component that visualizes the sequencing reads
and molecules processed at each stage of the sequencing pipeline.
}
