% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_saturation}
\alias{component_sequencing_saturation}
\title{Create the component sequencing saturation}
\usage{
component_sequencing_saturation(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing QC metrics and sample aliases.}
}
\value{
A list containing plots and a table summarizing the sequencing saturation.
}
\description{
This function creates a component that visualizes the sequencing saturation
for each sample.
}
