% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_umi_bias}
\alias{component_umi_bias}
\title{Create the component for umi bias}
\usage{
component_umi_bias(object, sample_palette, heatmap_gradient, min_cells = 50)
}
\arguments{
\item{object}{A Seurat object containing the processed data.}

\item{sample_palette}{A named character vector specifying the colors for each sample.}

\item{heatmap_gradient}{A character vector specifying the colors for the heatmap gradient.}

\item{min_cells}{An integer specifying the minimum number of cells required to include a sample in the analysis.}
}
\value{
A list of ggplot objects:
\itemize{
\item \code{p_degree_bias} : violin plot of the umi1 and umi2 degrees
\item \code{p_protein_bias} : heatmap of the correlation in abundance for each protein between umi1 and umi2
}
}
\description{
This function creates a plot the average umi1 and umi2 degree across samples.
}
