% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{convert_png_to_webp}
\alias{convert_png_to_webp}
\title{Convert PNG images to WebP format}
\usage{
convert_png_to_webp(filepath, options)
}
\arguments{
\item{filepath}{The path to the PNG file that knitr has just written.}

\item{options}{The options passed to the knitr hook.}
}
\value{
The path to the WebP file, which knitr will use for embedding in the document.
}
\description{
This function is used as a hook in knitr to convert PNG images generated during the rendering of a Quarto document
to WebP format. It reads the PNG file, converts it to WebP with specified quality, and then removes the original
PNG file.
}
