% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{create_sample_palette}
\alias{create_sample_palette}
\title{Create sample palette}
\usage{
create_sample_palette(conditions)
}
\arguments{
\item{conditions}{Character vector of conditions.}
}
\value{
Character vectors of color hex codes.
}
\description{
Create a sample palette based on the number of unique conditions in the experiment.
}
