% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\docType{data}
\name{displayed_cell_types}
\alias{displayed_cell_types}
\title{Cell types to display}
\format{
An object of class \code{character} of length 5.
}
\usage{
displayed_cell_types
}
\description{
This vector contains the cell types to be displayed in the ES, when filtering is applied.
}
\keyword{datasets}
