% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{.extract_legend}
\alias{.extract_legend}
\title{Extract Legend from a ggplot}
\usage{
.extract_legend(
  plot,
  legend_position = "bottom",
  plot_height = 5,
  justify_top = TRUE
)
}
\arguments{
\item{plot}{A ggplot object from which to extract the legend.}

\item{legend_position}{The position of the legend to extract (default is "bottom").}

\item{plot_height}{The height of the plot in inches (default is 5).}

\item{justify_top}{Logical indicating whether to justify the legend to the top (default is TRUE).}
}
\value{
A ggplot object containing only the legend.
}
\description{
Extracts the legend from a ggplot object and returns it as a separate ggplot object.
}
