% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{draw_quantiles}
\alias{draw_quantiles}
\title{Draw quantiles on a violin plot}
\usage{
draw_quantiles(
  p,
  draw_quantiles = 0.5,
  linewidth = 1,
  color = "gray20",
  facet_var = NULL,
  ...
)
}
\arguments{
\item{p}{A ggplot object representing the violin plot.}

\item{draw_quantiles}{A numeric vector of quantiles to draw (default is 0.5, which represents the median).}

\item{linewidth}{The width of the quantile lines (default is 1).}

\item{color}{The color of the quantile lines (default is "gray20").}

\item{facet_var}{An optional variable to facet the plot by (default is NULL).}

\item{...}{Additional arguments passed to \code{geom_segment}.}
}
\value{
A ggplot object with quantile lines added to the violin plot.
}
\description{
This function adds quantile lines to a violin plot created with ggplot2.
It calculates the quantiles for each violin and draws horizontal lines at those quantiles.
}
