% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_proximity_scores}
\alias{filter_proximity_scores}
\title{Filter Proximity Scores}
\usage{
filter_proximity_scores(pg_data_processed, params, sample_levels = NULL)
}
\arguments{
\item{pg_data_processed}{A Seurat object.}

\item{params}{A list of parameters for filtering, including control markers.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A data frame of filtered proximity scores with additional metadata.
}
\description{
This function filters the proximity scores from the \code{pg_data_processed} object.
}
