% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{format_with_info}
\alias{format_with_info}
\title{Format content with an info icon and tooltip}
\usage{
format_with_info(content, description)
}
\arguments{
\item{content}{A string containing the content to be formatted.}

\item{description}{A string containing the description to be displayed in the tooltip.}
}
\value{
A formatted string with an info icon and tooltip.
}
\description{
This function formats content with an info icon that displays a tooltip with additional information when hovered over.
}
