% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_coreness_data}
\alias{get_coreness_data}
\title{Get coreness data}
\usage{
get_coreness_data(object)
}
\arguments{
\item{object}{A Seurat object containing the coreness data.'}
}
\value{
A list containing the coreness data, component summary, and sample summary.
}
\description{
This function retrieves coreness data from a Seurat object and summarizes it by sample and component.
}
