% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_crossing_edges}
\alias{get_crossing_edges}
\title{Get crossing edges data}
\usage{
get_crossing_edges(sample_qc_metrics)
}
\arguments{
\item{sample_qc_metrics}{A list containing quality control metrics for each sample.}
}
\value{
A tibble containing the crossing edges data for each sample.
}
\description{
This function retrieves the crossing edges data from the sample quality control metrics.
}
