% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_degree_distribution}
\alias{get_degree_distribution}
\title{Get degree distribution data}
\usage{
get_degree_distribution(sample_qc_metrics)
}
\arguments{
\item{sample_qc_metrics}{A list containing quality control metrics for each sample.}
}
\value{
A tibble containing the degree distribution data for each sample.
}
\description{
This function retrieves degree distribution data from the sample quality control metrics.
}
