% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_denoising_data}
\alias{get_denoising_data}
\title{Get denoising data}
\usage{
get_denoising_data(sample_qc_metrics)
}
\arguments{
\item{sample_qc_metrics}{A list containing quality control metrics for each sample.}
}
\value{
A tibble containing the denoising data for each sample.
}
\description{
This function retrieves the denoising data from the sample quality control metrics.
}
