% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_qc_metrics}
\alias{get_qc_metrics}
\title{Get quality control metrics for samples}
\usage{
get_qc_metrics(object, sample_qc_metrics, sample_sheet)
}
\arguments{
\item{object}{A Seurat object containing the data.}

\item{sample_qc_metrics}{A list containing quality control metrics for each sample.}

\item{sample_sheet}{A data frame containing sample metadata, including \code{sample_alias}.}
}
\value{
A list containing various quality control metrics for each sample.
}
\description{
This function retrieves quality control metrics for samples from the provided sample QC metrics and sample sheet.
}
