% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_read_stats}
\alias{get_read_stats}
\title{Get read statistics for each sample}
\usage{
get_read_stats(object)
}
\arguments{
\item{object}{A Seurat object containing the data.}
}
\value{
A tibble summarizing read statistics for each sample.
}
\description{
This function retrieves read statistics for each sample from a Seurat object.
}
