% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_seq_saturation}
\alias{get_seq_saturation}
\title{Get sequencing saturation metrics}
\usage{
get_seq_saturation(object, sample_qc_metrics)
}
\arguments{
\item{object}{A Seurat object.}

\item{sample_qc_metrics}{A list containing quality control metrics for each sample.}
}
\value{
A tibble containing sequencing saturation metrics for each sample.
}
\description{
This function calculates sequencing saturation metrics for each sample based on the provided quality control metrics.
}
