% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_table.R
\name{get_top_markers}
\alias{get_top_markers}
\title{Get top markers for each group in the object}
\usage{
get_top_markers(object, group, n_markers = c(3, 5))
}
\arguments{
\item{object}{A Seurat object.}

\item{group}{A character string specifying the group variable in the object.}

\item{n_markers}{A numeric vector specifying the number of top markers to report for each group.}
}
\value{
A tibble with the group ID, top markers, and their fractions.
}
\description{
Calculate the top abundant markers for each group in the object and report the fraction of counts of the top markers
across all cells in the group.
}
