% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_registry.R
\name{list_es_data_workflows}
\alias{list_es_data_workflows}
\title{List registered Experiment Summary workflows}
\usage{
list_es_data_workflows()
}
\value{
A character vector of registered workflow identifiers.
}
\description{
List registered Experiment Summary workflows
}
