% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_pxl_data_list}
\alias{load_pxl_data_list}
\title{Load pxl files into a list}
\usage{
load_pxl_data_list(data_folder, data_files, sample_sheet)
}
\arguments{
\item{data_folder}{Path to the folder containing the data files.}

\item{data_files}{Paths to data files.}

\item{sample_sheet}{Sample sheet.}
}
\value{
A list of Seurat objects
}
\description{
Load pxl files into a list as specified by identified data paths and sample sheet.
}
