% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{merge_cell_types}
\alias{merge_cell_types}
\title{Merge cell types}
\usage{
merge_cell_types(
  object,
  celltype_col = "l1_annotation_summary",
  merged_col = "celltype",
  mapping = merged_cell_types
)
}
\arguments{
\item{object}{A Seurat object containing the cell type annotations in the metadata.}

\item{celltype_col}{The name of the column in the metadata that contains the original cell type annotations.
Default is "l1_annotation_summary".}

\item{merged_col}{The name of the new column to be added to the metadata that will contain the merged cell type
annotations. Default is "celltype".}

\item{mapping}{A named vector that maps original cell types to merged cell types. Default is the \code{merged_cell_types}
vector.}
}
\value{
A Seurat object with the merged cell types added to the metadata.
}
\description{
This function takes a Seurat object and merges the cell types based on the provided mapping. It adds a new column to
the metadata with the merged cell types.
}
