% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{order_cd_markers}
\alias{order_cd_markers}
\title{Order markers in a consistent way}
\usage{
order_cd_markers(markers, control_markers)
}
\arguments{
\item{markers}{A character vector of marker names.}

\item{control_markers}{A character vector of control marker names.}
}
\value{
A character vector of markers ordered as described above.
}
\description{
This function orders markers in a consistent way, with markers named "CD.*" first, non-CD markers second
in alphabetic order, and control markers last.
}
