% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\docType{data}
\name{pipeline_pool_stages}
\alias{pipeline_pool_stages}
\title{nf-core/pixelator pool stages}
\format{
An object of class \code{character} of length 4.
}
\usage{
pipeline_pool_stages
}
\description{
nf-core/pixelator pool stages
}
\keyword{datasets}
