% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\docType{data}
\name{pipeline_stages}
\alias{pipeline_stages}
\title{nf-core/pixelator stages}
\format{
An object of class \code{character} of length 9.
}
\usage{
pipeline_stages
}
\description{
nf-core/pixelator stages
}
\keyword{datasets}
