% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot Embedding}
\usage{
plot_embedding(
  object,
  plot_reduction = "pca",
  dims = 1:2,
  metavars = "seurat_clusters",
  pal = NULL,
  label = TRUE,
  plot_title = "",
  xaxis_title = TRUE,
  yaxis_title = TRUE,
  fix_aspect_ratio = TRUE,
  legend_position = "none",
  legend_cols = 2,
  extract_legend = FALSE,
  plot_height = 5
)
}
\arguments{
\item{object}{A Seurat object containing the embeddings to be plotted.}

\item{plot_reduction}{The name of the reduction to use for plotting (default is "pca").}

\item{dims}{A vector specifying the dimensions to plot (default is 1:2).}

\item{metavars}{A variable to color the points by (default is "seurat_clusters").}

\item{pal}{A color palette to use for the points (default is NULL).}

\item{label}{Logical indicating whether to label the points with their metadata (default is TRUE).}

\item{plot_title}{The title of the plot (default is an empty string).}

\item{xaxis_title}{Logical indicating whether to show the x-axis title (default is TRUE).}

\item{yaxis_title}{Logical indicating whether to show the y-axis title (default is TRUE).}

\item{fix_aspect_ratio}{Logical indicating whether to fix the aspect ratio of the plot (default is TRUE).}

\item{legend_position}{The position of the legend (default is "none").}

\item{legend_cols}{The number of columns in the legend (default is 2).}

\item{extract_legend}{Logical indicating whether to extract the legend from the plot (default is FALSE). If TRUE,
the function will return a list containing the plot and the legend separately.}

\item{plot_height}{The height of the plot in inches (default is 5).}
}
\value{
A ggplot object representing the embedding.
}
\description{
Plot a 2D embedding of a Seurat object, allowing for customization of the plot aesthetics.
}
