% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_embeddings_samplewise}
\alias{plot_embeddings_samplewise}
\title{Plot Embeddings Samplewise}
\usage{
plot_embeddings_samplewise(
  samples,
  object,
  plot_reduction = "pca",
  dims = 1:2,
  metavars = "seurat_clusters",
  pal = NULL,
  label = TRUE,
  plot_title = "",
  fix_aspect_ratio = TRUE,
  legend_position = "none",
  legend_cols = 2
)
}
\arguments{
\item{samples}{A vector of sample identifiers to plot.}

\item{object}{A Seurat object containing the embeddings to be plotted.}

\item{plot_reduction}{The name of the reduction to use for plotting (default is "pca").}

\item{dims}{A vector specifying the dimensions to plot (default is 1:2).}

\item{metavars}{A variable to color the points by (default is "seurat_clusters").}

\item{pal}{A color palette to use for the points (default is NULL).}

\item{label}{Logical indicating whether to label the points with their metadata (default is TRUE).}

\item{plot_title}{The title of the plot (default is an empty string).}

\item{fix_aspect_ratio}{Logical indicating whether to fix the aspect ratio of the plot (default is TRUE).}

\item{legend_position}{The position of the legend (default is "none").}

\item{legend_cols}{The number of columns in the legend (default is 2).}
}
\value{
A list of ggplot objects, each representing the embedding for a sample.
}
\description{
Plot embeddings for each sample in a Seurat object, allowing for customization of the plot aesthetics.
}
