% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_violin}
\alias{plot_violin}
\title{Make a violin plot}
\usage{
plot_violin(
  plot_data,
  x,
  y,
  fill = NULL,
  palette = NULL,
  title = NULL,
  subtitle = NULL,
  x_label = NULL,
  y_label = NULL,
  fill_label = NULL,
  summarize = TRUE,
  round = 1,
  use_pct = FALSE,
  use_1k = FALSE,
  expand = NULL,
  use_log10 = FALSE,
  alpha = 1,
  color = NA,
  draw_quantiles = 0.5,
  facet_var = NULL,
  use_grid = FALSE,
  use_jitter = FALSE,
  jitter_size = 0.1,
  jitter_alpha = 0.5,
  hline = NULL
)
}
\arguments{
\item{plot_data}{A data frame containing the data to plot.}

\item{x}{The variable to use for the x-axis.}

\item{y}{The variable to use for the y-axis.}

\item{fill}{Optional variable to fill color the violins.}

\item{palette}{Optional color palette to use for the fill variable.}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}

\item{fill_label}{The label for the fill variable.}

\item{summarize}{Logical indicating whether to show the median y value
above the 0.5 quantile.}

\item{round}{Number of decimal places to round the median y value.}

\item{use_pct}{Logical indicating whether to show the y-axis in percentage.}

\item{use_1k}{Logical indicating whether to divide median y value by 1,000.}

\item{expand}{Optional expansion factor for the y-axis.}

\item{use_log10}{Logical indicating whether to use a log10 scale for the y-axis.}

\item{alpha}{Transparency level for the violin fill.}

\item{color}{Color for the violin outline.}

\item{draw_quantiles}{Numeric vector of quantiles to draw on the violin plot (default is 0.5 for median).}

\item{facet_var}{Optional variable to facet the plot by.}

\item{use_grid}{Logical indicating whether to use a grid layout for the plot.}

\item{use_jitter}{Logical indicating whether to add quasirandom data points.}

\item{jitter_size}{Size of the quasirandom points.}

\item{jitter_alpha}{Transparency level for the quasirandom points (default 0.35).}

\item{hline}{Optional horizontal line to draw on the plot.}
}
\value{
A ggplot object representing the violin plot.
}
\description{
Make a violin plot
}
