% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{preferred_dimred_order}
\alias{preferred_dimred_order}
\title{Order reductions in a preferred way}
\usage{
preferred_dimred_order(reductions, keep_undefined = FALSE)
}
\arguments{
\item{reductions}{A character vector of reduction names.}

\item{keep_undefined}{A logical value indicating whether to keep reductions
that are not in the preferred order.}
}
\value{
A character vector of reductions ordered according to the preferred order.
}
\description{
This function orders dimensionality reductions in a preferred way, with
"umap" first, followed by "pca".
}
