% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_diagnostics_table}
\alias{print_diagnostics_table}
\title{Print Experiment Summary diagnostics}
\usage{
print_diagnostics_table(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object.}
}
\value{
A \code{datatables} HTML widget containing all diagnostics.
}
\description{
Prints all diagnostics recorded while building an \code{es_data} object as a
styled table.
}
