% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_params}
\alias{print_params}
\title{Print analysis parameters}
\usage{
print_params(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing analysis parameters.}
}
\value{
A printed table of parameters with their values and defaults.
}
\description{
Print the analysis parameters in a table format, comparing them to the default parameters.
}
