% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_pixelator_version}
\alias{print_pixelator_version}
\title{Print Pixelator Version table}
\usage{
print_pixelator_version(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing processed PXL data and the
effective samplesheet.}
}
\value{
A printed table of Pixelator version information.
}
\description{
Print the Pixelator version information from the pxl files.
}
