% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_session_info}
\alias{print_session_info}
\title{Print Session Info table}
\usage{
print_session_info()
}
\value{
A printed table of session info.
}
\description{
Print the session info in a table format.
}
