% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_proximity_scores}
\alias{process_proximity_scores}
\title{Process Proximity Scores}
\usage{
process_proximity_scores(pg_data_processed, proximity_scores)
}
\arguments{
\item{pg_data_processed}{A Seurat object containing processed data.}

\item{proximity_scores}{A data frame of proximity scores.}
}
\value{
A data frame.
}
\description{
This function processes the proximity scores to prepare them for plotting.
}
