% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{read_annotation_reference}
\alias{read_annotation_reference}
\title{Read cell annotation reference}
\usage{
read_annotation_reference()
}
\value{
A Seurat object containing the cell annotation and counts data.
}
\description{
This function reads the cell annotation reference data from the package's extdata directory.
}
