% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_abundance_anova}
\alias{run_abundance_anova}
\title{Run ANOVA for abundance data in a Seurat object}
\usage{
run_abundance_anova(
  object,
  vars,
  layer = "data",
  mc_cores = 1,
  p_adj_method = "bonferroni"
)
}
\arguments{
\item{object}{A Seurat object containing the abundance data.}

\item{vars}{A character vector of length 1 or 2 specifying the variables to use in the ANOVA.}

\item{layer}{A string specifying the layer of the Seurat object to use for the analysis (default is "data").}

\item{mc_cores}{An integer specifying the number of cores to use for parallel processing (default is 1).}

\item{p_adj_method}{A string specifying the method for adjusting p-values (default is "bonferroni").}
}
\value{
A data frame with ANOVA results for each marker, including adjusted p-values.
}
\description{
This function runs an ANOVA for abundance data in a Seurat object, allowing for parallel processing.
}
