% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_proximity_anova}
\alias{run_proximity_anova}
\title{Run ANOVA for proximity scores in a Seurat object}
\usage{
run_proximity_anova(
  proximity_scores,
  vars,
  proximity_score = "log2_ratio",
  mc_cores = 1,
  p_adj_method = "bonferroni"
)
}
\arguments{
\item{proximity_scores}{A data frame containing proximity scores.}

\item{vars}{A character vector of length 1 or 2 specifying the variables to use in the ANOVA.}

\item{proximity_score}{A string specifying the column name for proximity scores (default is "log2_ratio").}

\item{mc_cores}{An integer specifying the number of cores to use for parallel processing (default is 1).}

\item{p_adj_method}{A string specifying the method for adjusting p-values (default is "bonferroni").}
}
\value{
A data frame with ANOVA results for each marker, including adjusted p-values.
}
\description{
This function runs an ANOVA for proximity scores in a Seurat object, allowing for parallel processing.
}
