% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{section_intro}
\alias{section_intro}
\title{Render a section with introductory text}
\usage{
section_intro(title, text, level = 3)
}
\arguments{
\item{title}{A string representing the title of the section.}

\item{text}{A string representing the introductory text to be displayed}

\item{level}{An integer representing the heading level for the section
title (default is 3).}
}
\description{
This function is a helper for rendering a section with introductory text
in the ES report. It takes a title for the section, an optional level for
the heading (default is 3), and the text to be displayed. It then
prints the section heading and the introductory text in the appropriate
format for the ES report.
}
