% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_utils.R
\name{section_table}
\alias{section_table}
\title{Render a section with a table}
\usage{
section_table(table, title, level = 3)
}
\arguments{
\item{table}{A datatables object to be rendered in the section.}

\item{title}{A string representing the title of the section.}

\item{level}{An integer representing the heading level for the section
title (default is 3).}
}
\description{
This function is a helper for rendering a section with a
table in the ES report. It takes a datatables object, a title
for the section, and an optional level for the heading (default is 3).
It then prints the section heading and the table in the appropriate
format for the ES report.
}
