% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{set_sample_levels}
\alias{set_sample_levels}
\title{Set sample levels in a data frame}
\usage{
set_sample_levels(df, sample_levels = NULL)
}
\arguments{
\item{df}{A data frame containing a \code{sample_alias} column.}

\item{sample_levels}{A vector of sample levels to set for the \code{sample_alias} column (default is NULL).}
}
\value{
A data frame with the \code{sample_alias} column set to the specified levels.
}
\description{
This function sets the levels of the \code{sample_alias} factor in a data frame.
}
