% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{style_table}
\alias{style_table}
\title{Create a HTML styled table}
\usage{
style_table(
  df,
  pageLength = NULL,
  caption = NULL,
  interactive = TRUE,
  search = TRUE,
  lengthChange = FALSE,
  buttons = TRUE,
  escape = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame}

\item{pageLength}{An integer specifying the number of rows to display per page.
If NULL, it defaults to the number of rows in the data frame.}

\item{caption}{A string to be used as the table caption. If NULL, no caption is displayed.}

\item{interactive}{A boolean indicating whether the table should contains interactive
elements (default is TRUE).}

\item{search}{A boolean indicating whether to include a search box (default is TRUE).}

\item{lengthChange}{A boolean indicating whether to allow users to change the number
of rows displayed (default is FALSE).}

\item{buttons}{A boolean indicating whether to include download buttons for CSV and Excel
files (default is TRUE).}

\item{escape}{A boolean indicating whether to escape HTML characters in the table cells
(default is FALSE).}

\item{...}{Additional arguments passed to the \code{DT::datatable} function.}
}
\value{
A HTML table using the DT package.
}
\description{
Create a HTML styled table
}
