% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{summarize_colocalization_scores_per_sample}
\alias{summarize_colocalization_scores_per_sample}
\title{Summarize proximity scores per sample and condition}
\usage{
summarize_colocalization_scores_per_sample(proximity_scores, test_mode = FALSE)
}
\arguments{
\item{proximity_scores}{A data frame of proximity scores.}

\item{test_mode}{A logical indicating whether to run in test mode (default is FALSE).}
}
\value{
A data frame summarizing the proximity scores per sample and condition.
}
\description{
This function summarizes the proximity per sample and condition.
}
