% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ES.R
\name{test_ES}
\alias{test_ES}
\title{Runs a ridiculously minimal test of the pixelatorESinternal}
\usage{
test_ES(
  output_file = NULL,
  use_tiny_test = TRUE,
  params = NULL,
  view_ES = FALSE,
  ...
)
}
\arguments{
\item{output_file}{An optional specifying the output file where the
rendered document will be saved.}

\item{use_tiny_test}{A boolean indicating whether to use a tiny test dataset
with customized parameters (default is TRUE).}

\item{params}{A list of parameters to override the default settings.}

\item{view_ES}{A boolean indicating whether to open the rendered HTML file}

\item{...}{Additional arguments passed to the \code{quarto_render} function.}
}
\value{
Nothing. Only used for its side effects
}
\description{
This function is intended for development to quickly test if
new features break. It uses a combinations of json files from
\code{pixelatorESinternal} and PXL files from \code{pixelatorR}. The PXL
files only contain 5 cells and the data is therefore not useful
to evaluate most plots.
}
\details{
It is possible to test with a larger dataset by providing a custom
\code{params} list where the \code{sample_sheet} and \code{data_folder} are changed.
Note that you should also set \code{use_tiny_test = FALSE} to use more
sutiable defaults for the test in this case.
}
\section{Quarto path}{

It's important to make sure that quarto is isntalled and that \code{quarto_render}
uses the correct quarto path. The easiest way to set this is by using:
\code{Sys.setenv(QUARTO_PATH = "<path to quarto binary>")}
}

