% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_violin}
\alias{theme_violin}
\title{Theme for violin plots}
\usage{
theme_violin(base_size = 10)
}
\arguments{
\item{base_size}{Base font size for the theme.}
}
\value{
A ggplot2 theme object.
}
\description{
Theme for violin plots
}
