% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{yes_no_palette}
\alias{yes_no_palette}
\title{Yes/No gradient}
\format{
An object of class \code{character} of length 2.
}
\usage{
yes_no_palette
}
\description{
Yes/No gradient
}
\keyword{datasets}
