% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{complete_proximity_scores}
\alias{complete_proximity_scores}
\title{Complete Proximity Scores}
\usage{
complete_proximity_scores(proximity_scores, only_self = TRUE)
}
\arguments{
\item{proximity_scores}{A data frame of proximity scores.}

\item{only_self}{A boolean indicating whether to filter for self-comparisons only (default is TRUE).}
}
\value{
A data frame of completed proximity scores with missing values filled.
}
\description{
This function completes the proximity scores by ensuring that each marker pair has a score for each sample component,
filling missing values with 0.
}
