% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_abundance_per_marker}
\alias{component_abundance_per_marker}
\title{Create the component for abundance per marker}
\usage{
component_abundance_per_marker(
  object,
  params,
  sample_palette,
  per_celltype = TRUE,
  sample_levels = NULL,
  test_mode = FALSE
)
}
\arguments{
\item{object}{A processed data object containing marker abundance data.}

\item{params}{A list of parameters, including control markers.}

\item{sample_palette}{A color palette for the samples.}

\item{per_celltype}{A boolean indicating whether to add a plot per cell type
(default is TRUE).}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots for each marker.
}
\description{
This function creates plots visualizing the abundance of each marker
across samples and, optionally, across cell types.
}
