% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_annotation}
\alias{component_annotation}
\title{Create the component annotation summary}
\usage{
component_annotation(object, heatmap_gradient, reduction = NULL)
}
\arguments{
\item{object}{A Seurat object containing the processed data.}

\item{heatmap_gradient}{A color gradient for the heatmap.}

\item{reduction}{Optional character specifying the reduction to use. If NULL, the first preferred reduction is used.}
}
\value{
A list of ggplot objects visualizing the l1 and l2 annotation summaries.
}
\description{
This function creates a component that visualizes the annotation summary
for l1 and l2 annotations in the specified reduction.
}
