% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_cell_recovery}
\alias{component_cell_recovery}
\title{Create the component cell recovery}
\usage{
component_cell_recovery(object, sample_qc_metrics, sample_levels = NULL)
}
\arguments{
\item{object}{A Seurat object containing the sample data.}

\item{sample_qc_metrics}{A list of sample QC metrics.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list containing a plot and a table summarizing the cell recovery.
}
\description{
This function creates a component that visualizes the cell recovery
after filtering components based on size.
}
