% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_control_markers}
\alias{component_control_markers}
\title{Create the component for isotype control markers}
\usage{
component_control_markers(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing processed PXL data.}
}
\value{
A list containing two violin plots and a summary table.
}
\description{
This function creates a list with two plots and a summary table for
isotype control marker fraction and counts.
}
