% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_control_markers}
\alias{component_control_markers}
\title{Create the component for control markers}
\usage{
component_control_markers(pg_data)
}
\arguments{
\item{pg_data}{A Seurat object containing the data to be plotted.}
}
\value{
A list containing two violin plots and a summary table.
}
\description{
This function creates a list with two plots and a summary table.
}
