% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_coreness}
\alias{component_coreness}
\title{Create the component coreness}
\usage{
component_coreness(qc_metrics_tables)
}
\arguments{
\item{qc_metrics_tables}{A list of QC metrics tables containing coreness data.}
}
\value{
A list containing three plots and a table summarizing the coreness metrics.
}
\description{
This function creates a component that visualizes the k-coreness of nodes in each component.
}
