% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_crossing_edges}
\alias{component_crossing_edges}
\title{Create the component crossing edges}
\usage{
component_crossing_edges(qc_metrics_tables, sample_levels = NULL)
}
\arguments{
\item{qc_metrics_tables}{A list of QC metrics tables.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list containing a plot and a table summarizing the crossing edges.
}
\description{
This function creates a component that visualizes the percentage of crossing edges
for each sample, both in the initial and refinement stages.
}
