% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_molecule_plot}
\alias{component_molecule_plot}
\title{Create the component for molecule plot}
\usage{
component_molecule_plot(pg_data, sample_palette)
}
\arguments{
\item{pg_data}{A Seurat object containing the data to be plotted.}

\item{sample_palette}{A named vector of colors for the samples.}
}
\value{
A ggplot object representing the violin plot of molecule counts.
}
\description{
This function generates a violin plot showing the distribution of the number
of molecules per sample, with a horizontal line indicating the molecule count cutoff.
}
