% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_node_degree}
\alias{component_node_degree}
\title{Create the component node degree}
\usage{
component_node_degree(object, sample_levels = NULL)
}
\arguments{
\item{object}{A Seurat object containing the sample data.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list containing a plot and a table summarizing the mean node degree.
}
\description{
This function creates a component that visualizes the mean node degree for A and B nodes
for each sample.
}
