% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_node_edge_count}
\alias{component_node_edge_count}
\title{Create the component node and edge count}
\usage{
component_node_edge_count(sample_qc_metrics, sample_levels = NULL)
}
\arguments{
\item{sample_qc_metrics}{A list of sample QC metrics.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list containing two plots and a table summarizing the node and edge counts.
}
\description{
This function creates a component that visualizes the number of nodes and edges
before and after cell recovery.
}
