% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_heatmap_celltype}
\alias{component_proximity_heatmap_celltype}
\title{Create the component for proximity heatmaps per celltype and condition}
\usage{
component_proximity_heatmap_celltype(
  processed_data,
  heatmap_gradient,
  proximity_score
)
}
\arguments{
\item{processed_data}{A processed data object containing marker abundance data.}

\item{heatmap_gradient}{A color palette for the heatmaps.}

\item{proximity_score}{One of "join_count_z" or "log2_ratio".}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the proximity scores as heatmaps per celltype and condition.
}
