% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_heatmap_sample}
\alias{component_proximity_heatmap_sample}
\title{Create the component for proximity heatmaps per sample and condition}
\usage{
component_proximity_heatmap_sample(
  proximity_scores,
  heatmap_gradient,
  proximity_score = "log2_ratio",
  test_mode = FALSE
)
}
\arguments{
\item{proximity_scores}{A data frame containing proximity scores for different markers.}

\item{heatmap_gradient}{A color palette for the heatmaps.}

\item{proximity_score}{One of "join_count_z" or "log2_ratio".}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the proximity scores as heatmaps per sample and condition.
}
