% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_per_marker}
\alias{component_proximity_per_marker}
\title{Create the component for proximity per marker}
\usage{
component_proximity_per_marker(
  es_data,
  sample_palette,
  per_celltype = TRUE,
  test_mode = FALSE
)
}
\arguments{
\item{es_data}{An \code{es_data} object containing proximity scores and sample aliases.}

\item{sample_palette}{A color palette for the samples.}

\item{per_celltype}{A boolean indicating whether to add a plot per cell type (default is TRUE).}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots for each marker.
}
\description{
This function creates plots visualizing the proximity scores for each marker
across different samples and conditions. Every component is shown for every
plotted marker, with proximity scores that were filtered out set to 0.
}
