% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_per_marker}
\alias{component_proximity_per_marker}
\title{Create the component for proximity per marker}
\usage{
component_proximity_per_marker(
  processed_data,
  proximity_score,
  sample_palette,
  test_mode = FALSE
)
}
\arguments{
\item{processed_data}{A processed data object containing proximity scores.}

\item{proximity_score}{One of "join_count_z" or "log2_ratio".}

\item{sample_palette}{A color palette for the samples.}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots for each marker.
}
\description{
This function creates plots visualizing the proximity scores for each marker
across different samples and conditions.
}
