% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_selected}
\alias{component_proximity_selected}
\title{Create the component for proximity Z scores}
\usage{
component_proximity_selected(
  object,
  proximity_scores,
  sample_palette,
  proximity_score = "log2_ratio",
  selected_contrasts = TRUE,
  sample_levels = NULL,
  test_mode = FALSE
)
}
\arguments{
\item{object}{A Seurat object containing the sample data.}

\item{sample_palette}{A color palette for the samples.}

\item{proximity_score}{One of "join_count_z" or "log2_ratio".}

\item{selected_contrasts}{A boolean indicating whether to filter for selected contrasts (default is TRUE).}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}

\item{processed_data}{A processed data object containing marker abundance data.}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the proximity Z scores for selected
contrasts.
}
