% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_proximity_selected}
\alias{component_proximity_selected}
\title{Create the component for proximity Z scores}
\usage{
component_proximity_selected(
  processed_data,
  sample_palette,
  proximity_score,
  test_mode = FALSE
)
}
\arguments{
\item{processed_data}{A processed data object containing marker abundance data.}

\item{sample_palette}{A color palette for the samples.}

\item{proximity_score}{One of "join_count_z" or "log2_ratio".}

\item{test_mode}{A boolean indicating whether to run in test mode (default is FALSE).}
}
\value{
A list containing plots.
}
\description{
This function creates plots visualizing the proximity Z scores for selected
contrasts.
}
