% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_reads_and_molecules}
\alias{component_sequencing_reads_and_molecules}
\title{Create the component sequencing reads and molecules component}
\usage{
component_sequencing_reads_and_molecules(
  sample_qc_metrics,
  sample_levels = NULL
)
}
\arguments{
\item{sample_qc_metrics}{A list of sample QC metrics.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list containing plots and a table.
}
\description{
This function creates a component that visualizes the sequencing reads
and molecules processed at each stage of the sequencing pipeline.
}
