% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_reads_per_cell}
\alias{component_sequencing_reads_per_cell}
\title{Create the component reads per cell}
\usage{
component_sequencing_reads_per_cell(object)
}
\arguments{
\item{object}{A Seurat object.}
}
\value{
A list containing a plot and a table.
}
\description{
This function creates a component that visualizes the reads per cell per sample.
}
