% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_saturation}
\alias{component_sequencing_saturation}
\title{Create the component sequencing saturation}
\usage{
component_sequencing_saturation(qc_metrics_tables, sample_levels = NULL)
}
\arguments{
\item{qc_metrics_tables}{A list of QC metrics tables.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}
}
\value{
A list containing plots and a table summarizing the sequencing saturation.
}
\description{
This function creates a component that visualizes the sequencing saturation
for each sample.
}
