% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_saturation_curve}
\alias{component_sequencing_saturation_curve}
\title{Create the component sequencing saturation curve}
\usage{
component_sequencing_saturation_curve(es_data, seqsat_comps = 100L)
}
\arguments{
\item{es_data}{An \code{es_data} object containing processed PXL data, discovered
data files, the effective samplesheet, and sample aliases.}

\item{seqsat_comps}{The number of components to consider for sequencing saturation.}
}
\value{
A list containing plots showing the sequencing saturation curve for
each sample.
}
\description{
This function creates a component that visualizes the sequencing saturation
curve for each sample.
}
