% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component_sequencing_saturation_curve}
\alias{component_sequencing_saturation_curve}
\title{Create the component sequencing saturation curve}
\usage{
component_sequencing_saturation_curve(
  object,
  data_files,
  seqsat_comps = 5L,
  sample_levels = NULL,
  mc_cores = 1
)
}
\arguments{
\item{object}{A Seurat object containing the sample data.}

\item{data_files}{A data frame containing the sample aliases and corresponding file names.}

\item{seqsat_comps}{The number of components to consider for sequencing saturation.}

\item{sample_levels}{Optional vector of sample levels to order the samples in the plots.}

\item{mc_cores}{The number of cores to use for parallel processing. Default is 1, meaning
sequential processing.}
}
\value{
A list containing plots showing the sequencing saturation curve for
each sample.
}
\description{
This function creates a component that visualizes the sequencing saturation
curve for each sample.
}
