% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{Default parameters for the ES}
\format{
An object of class \code{list} of length 10.
}
\usage{
default_params
}
\description{
Default parameters for the ES
}
\keyword{datasets}
