% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{downsample_data}
\alias{downsample_data}
\title{Downsample data to a specified number of cells and markers}
\usage{
downsample_data(pg_data, control_markers = NULL, n_cells = 50, n_markers = 20)
}
\arguments{
\item{pg_data}{A Seurat object containing the data to be downsampled.}

\item{control_markers}{A character vector of control markers to always include in the downsampled data.}

\item{n_cells}{An integer specifying the number of cells to keep in each sample.}

\item{n_markers}{An integer specifying the total number of markers to keep in the downsampled data.}
}
\value{
A downsampled Seurat object with the specified number of cells and markers.
}
\description{
Downsamples the data to a specified number of cells and markers, ensuring that control markers are always included.
}
