% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{extract_sample_qc_metrics}
\alias{extract_sample_qc_metrics}
\title{Extract sample QC metrics from a list of sample QC data}
\usage{
extract_sample_qc_metrics(sample_qc_metrics, vars, stage = "graph")
}
\arguments{
\item{sample_qc_metrics}{A list of sample QC data, where each element is a named list
containing QC metrics for a sample.}

\item{vars}{A character vector of variable names to extract from the sample QC data.}

\item{stage}{A character string specifying the stage from which to extract the metrics.}
}
\value{
A tibble with the sample alias and the specified metrics.
}
\description{
This function extracts specific quality control metrics from a list of sample
QC data (as generated by \code{read_qc_files}). It returns a tibble with the
sample alias and the specified metrics.
}
