% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{find_stage}
\alias{find_stage}
\title{Find pixelator stage of file}
\usage{
find_stage(filepath)
}
\arguments{
\item{filepath}{A file path}
}
\value{
The name of the stage it comes from
}
\description{
Determine which pixelator stage a file comes from
}
