% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_file_paths}
\alias{get_file_paths}
\title{Get file paths for data and QC files}
\usage{
get_file_paths(
  data_folder = NULL,
  file_paths = NULL,
  sample_aliases,
  allow_unknown = FALSE
)
}
\arguments{
\item{data_folder}{A character string specifying the path to the data folder.}

\item{file_paths}{A character vector with file paths including all data files.}

\item{sample_aliases}{A named character vector mapping sample aliases to their actual names.}

\item{allow_unknown}{A logical indicating whether to allow files from unknown stages (default: FALSE).}
}
\value{
A list containing two data frames: \code{data_files} and \code{qc_files}.
}
\description{
Retrieve file paths for data and QC files from a specified data folder.
}
