% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get test data}
\usage{
get_test_data(concatenate = TRUE, type = c("default", "hashing"))
}
\arguments{
\item{concatenate}{A logical indicating whether to concatenate the data 6 times (default is TRUE).}

\item{type}{A character string specifying the type of test data to generate. Options are "default"
(the standard test data) and "hashing" (test data for hashing experiments}
}
\value{
A Seurat object containing test data with normalized and scaled data, PCA results, and merged layers.
}
\description{
Generates a minimal Seurat object for testing purposes.
}
