% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get test data}
\usage{
get_test_data()
}
\value{
A Seurat object containing test data with normalized and scaled data, PCA results, and merged layers.
}
\description{
Generates a minimal Seurat object for testing purposes.
}
