% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_test_qc_metrics}
\alias{get_test_qc_metrics}
\title{Get test QC metrics}
\usage{
get_test_qc_metrics(type = c("default", "hashing"))
}
\arguments{
\item{type}{A character string specifying the type of test data to use. Options are "default"
(the standard test data) and "hashing" (test data for hashing experiments).}
}
\value{
A list of QC metrics for each sample, where each element is a named list of metrics.
}
\description{
Reads the test samplesheet and retrieves QC metrics from the test data folder.
}
