% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_test_qc_metrics}
\alias{get_test_qc_metrics}
\title{Get test QC metrics}
\usage{
get_test_qc_metrics()
}
\value{
A list of QC metrics for each sample, where each element is a named list of metrics.
}
\description{
Reads the test samplesheet and retrieves QC metrics from the test data folder.
}
