% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge data from multiple samples into a single Seurat object}
\usage{
merge_data(pg_data, sample_sheet)
}
\arguments{
\item{pg_data}{A list of Seurat objects, each representing a sample.}

\item{sample_sheet}{A data frame containing sample metadata, including \code{sample_alias} and \code{condition}.}
}
\value{
A merged Seurat object with added metadata.
}
\description{
Merges data from list of multiple samples into a single Seurat object, adding metadata and ranks based on UMI counts.
}
