% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\docType{data}
\name{merged_cell_types}
\alias{merged_cell_types}
\title{Cell type merging}
\format{
An object of class \code{character} of length 6.
}
\usage{
merged_cell_types
}
\description{
This vector contains the mapping of original cell types to merged cell types, which is used by the ES to merge cell
types for display purposes. This allows for a more concise and interpretable visualization.
}
\keyword{datasets}
