% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_utils.R
\name{order_sample_alias_factors}
\alias{order_sample_alias_factors}
\title{Order sample alias factors in a data frame}
\usage{
order_sample_alias_factors(object, levels)
}
\arguments{
\item{object}{A data frame containing a \code{sample_alias} column.}

\item{levels}{A character vector specifying the levels to order the \code{sample_alias} factor.}
}
\value{
A data frame with the \code{sample_alias} factor ordered according to the specified levels.
}
\description{
This function orders the \code{sample_alias} factor in a data frame according to the specified levels.
}
