% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelatorES-package.R
\docType{package}
\name{pixelatorES-package}
\alias{pixelatorES}
\alias{pixelatorES-package}
\title{pixelatorES: Experiment Summary}
\description{
The Experiment Summary (ES) generates a self-contained `.html` report summarizing a multi-sample experiment, including QC metrics, statistics, and visualizations.
}
\author{
\strong{Maintainer}: Max Karlsson \email{max.karlsson@pixelgen.com} (\href{https://orcid.org/0000-0002-7000-4416}{ORCID})

Authors:
\itemize{
  \item Ludvig Larsson \email{ludvig.larsson@pixelgen.com} (\href{https://orcid.org/0000-0003-4209-2911}{ORCID})
}

}
\keyword{internal}
