% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_data.R
\name{print.es_data}
\alias{print.es_data}
\title{Print an \code{es_data} object}
\usage{
\method{print}{es_data}(x, ...)
}
\arguments{
\item{x}{An \code{es_data} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints the workflow and a concise description of the contents of every
\code{es_data} slot.
}
