% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_metadata_table}
\alias{print_metadata_table}
\title{Print metadata table}
\usage{
print_metadata_table(es_data)
}
\arguments{
\item{es_data}{An \code{es_data} object containing the samplesheet and processed
PXL data.}
}
\value{
A printed table of sample metadata.
}
\description{
Print the experiment meta data in a table format.
}
