% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_metadata_table}
\alias{print_metadata_table}
\title{Print metadata table}
\usage{
print_metadata_table(sample_sheet)
}
\arguments{
\item{sample_sheet}{A sample sheet.}
}
\value{
A printed table of sample metadata.
}
\description{
Print the experiment meta data in a table format.
}
