% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_params}
\alias{print_params}
\title{Print analysis parameters}
\usage{
print_params(params)
}
\arguments{
\item{params}{A list of parameters used in the analysis.}
}
\value{
A printed table of parameters with their values and defaults.
}
\description{
Print the analysis parameters in a table format, comparing them to the default parameters.
}
