% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print_pixelator_version}
\alias{print_pixelator_version}
\title{Print Pixelator Version table}
\usage{
print_pixelator_version(object, sample_sheet)
}
\arguments{
\item{object}{A Seurat object.}

\item{sample_sheet}{A sample sheet.}
}
\value{
A printed table of Pixelator version information.
}
\description{
Print the Pixelator version information from the pxl files.
}
