% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_qc_files}
\alias{read_qc_files}
\title{Read QC files and return metrics}
\usage{
read_qc_files(qc_files, sample_sheet)
}
\arguments{
\item{qc_files}{A data frame containing the file paths of json QC files.}

\item{sample_sheet}{A data frame containing sample metadata, including \code{sample_alias}.}
}
\value{
A list of QC metrics for each sample, where each element is a named list of metrics.
}
\description{
Reads QC files and returns a list of metrics for each sample.
}
