% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_samplesheet}
\alias{read_samplesheet}
\title{Read a sample sheet from a CSV file}
\usage{
read_samplesheet(filepath, additional_columns = NULL)
}
\arguments{
\item{filepath}{A character string specifying the path to the sample sheet CSV file.}
}
\value{
A tibble with columns \code{sample}, \code{sample_alias}, \code{condition}. Columns \code{pool}, \code{lot_role}, and \code{kit_lot_id} are
included if present in the CSV file.
}
\description{
Reads a sample sheet from a CSV file and returns it as a tibble.
}
