% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{tabset_figure_table}
\alias{tabset_figure_table}
\title{Tabset a figure and a table}
\usage{
tabset_figure_table(figure, table, level = 2, mode = "tabset")
}
\arguments{
\item{figure}{A ggplot object representing the figure to be tabsetted.}

\item{table}{A table to be tabsetted alongside the figure.}

\item{level}{Integer indicating the header level for the tabset title (default is 2).}

\item{mode}{Character indicating the mode of the tabset ("tabset" or "title").}
}
\value{
A formatted tabset containing the figure and table.
}
\description{
Tabset a figure and a table at a given header level. After using this
function, you should close the tabset with \code{close_tabset()}.
}
