% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{tabset_nested_plotlist}
\alias{tabset_nested_plotlist}
\title{Tabset nested plots}
\usage{
tabset_nested_plotlist(plots, level = 2, close = TRUE)
}
\arguments{
\item{plots}{List of plots or list of lists of plots to be tabsetted.}

\item{level}{Integer indicating the header level for the tabset title (default is 2).}

\item{close}{Logical indicating whether to close the tabset after plotting (default is TRUE).}
}
\value{
A formatted tabset of plots.
}
\description{
Tabset a list of plots, which may contain nested lists of plots, at a given header level.
}
