% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{tabset_plotlist}
\alias{tabset_plotlist}
\title{Tabset plots}
\usage{
tabset_plotlist(plots, level = 2)
}
\arguments{
\item{plots}{List of plots to be tabsetted.}

\item{level}{Integer indicating the header level for the tabset title (default is 2).}
}
\value{
A formatted tabset of plots.
}
\description{
Tabset a list of plots at a given header level.
}
