% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{test_data_folder}
\alias{test_data_folder}
\title{Get the path to the test data folder}
\usage{
test_data_folder(type = c("default", "hashing"), package = "pixelatorES")
}
\arguments{
\item{type}{A character string specifying the type of test data folder to return. Options are "default"
(the standard test data folder) and "hashing" (a test data folder for hashing experiments). Default is "default".}

\item{package}{A character string specifying the name of the package where the test data folder is located.
Default is "pixelatorES".}
}
\value{
A character string with the path to the test data folder.
}
\description{
Returns the path to the test data folder containing QC JSON files included in the package.
}
