% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{test_data_folder}
\alias{test_data_folder}
\title{Get the path to the test data folder}
\usage{
test_data_folder()
}
\value{
A character string with the path to the test data folder.
}
\description{
Returns the path to the test data folder containing QC JSON files included in the package.
}
