% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{test_samplesheet}
\alias{test_samplesheet}
\title{Get the path to the test samplesheet}
\usage{
test_samplesheet()
}
\value{
A character string with the path to the test samplesheet CSV file.
}
\description{
Returns the path to the test samplesheet CSV file included in the package.
}
