% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{title_plotlist}
\alias{title_plotlist}
\title{Header wrap plots}
\usage{
title_plotlist(plots, level = 2)
}
\arguments{
\item{plots}{List of plots to be wrapped.}

\item{level}{Integer indicating the header level for the headers (default is 2).}
}
\value{
A formatted list of plots with headers.
}
\description{
Wraps a list of plots with headers for each plot.
}
